Added textbook examples from Java Foundations 5E
Added textbook examples from Chapter 2 including an
example with the backspace escape sequence character,
a simple Scanner echo program, and a GasMileage program.

Errors for the GasMileage program include ability for
users to enter erroneous data which will lead to an exceptions
error. Specifically InputMismatchException. This is not covered
in the text yet.