Added a basic tech tree program
The program lets you choose 1 of 3 options. Once you choose one, the
"technology" changes to a new one that you can then choose. If you
get to the end of the tech tree, the selection changes to "XXXXX".
Once you get to the end of all three tech trees, the program exits.