Created IntDivideTest and CircleArea
First program takes in two integers by a user, the dividend and
the divisor, does integer division on it so it gets a quotient and
a remainder, puts that to the System output, and then finds the
entered integer by multiplying the quotient by the divisor and
adding the remainder to it (quotient * divisor + remainder).

CircleArea takes in a radius and calculates its area from the user
input.

Neither program handles exceptions at this point.