Created Dice Game Program
This dice game program has the computer throw two dice and the user
must beat that dice value and can choose to do so using uniform
distribution (one 11-sided die) or normal distribution (two 6-sided die)

If the value is less than 7, it is better to use the two dice and if it
is greater than 7, your odds are better with one die.

If the computer or player wins 10 rounds, they win the game and the
program exits.